2010-12-10  Hans Wennborg  <hans@chromium.org>
        Reviewed by Jeremy Orlow.

        IndexedDB: Numeric keys are floats.
        https://bugs.webkit.org/show_bug.cgi?id=50674

        Update layout tests to use floating-point values for keys.

        * storage/indexeddb/keyrange-expected.txt:
        * storage/indexeddb/keyrange.html:
        Make sure creating the keys works.
        * storage/indexeddb/objectstore-cursor-expected.txt:
        * storage/indexeddb/objectstore-cursor.html:
        Make sure retrieving the keys from the DB objectstore works.
2010-12-10  Hans Wennborg  <hans@chromium.org>

        Reviewed by Jeremy Orlow.

        IndexedDB: Numeric keys are floats.
        https://bugs.webkit.org/show_bug.cgi?id=50674

        Use floating point to represent numeric keys,
        add version meta data to the SQLite db,
        and migrate object stores that use integers.

        * bindings/v8/IDBBindingUtilities.cpp:
        (WebCore::createIDBKeyFromValue):
        * bindings/v8/custom/V8IDBKeyCustom.cpp:
        (WebCore::toV8):
        * storage/IDBFactoryBackendImpl.cpp:
        (WebCore::createTables):
        (WebCore::migrateDatabase):
        (WebCore::IDBFactoryBackendImpl::open):
        * storage/IDBKey.cpp:
        (WebCore::IDBKey::IDBKey):
        (WebCore::IDBKey::fromQuery):
        (WebCore::IDBKey::bind):
        (WebCore::IDBKey::bindWithNulls):
        * storage/IDBKey.h:
        (WebCore::IDBKey::create):
        (WebCore::IDBKey::number):
2010-12-10  Hans Wennborg  <hans@chromium.org>

        Reviewed by Jeremy Orlow.

        IndexedDB: Numeric keys are floats.
        https://bugs.webkit.org/show_bug.cgi?id=50674

        Represent numeric keys as floating point values.

        * public/WebIDBKey.h:
        (WebKit::WebIDBKey::WebIDBKey):
        * src/WebIDBKey.cpp:
        (WebKit::WebIDBKey::assign):
        (WebKit::WebIDBKey::number):

Canonical link: https://commits.webkit.org/64104@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@73697 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/storage/IDBKey.h
@@ -43,7 +43,7 @@ public:
     {
         return adoptRef(new IDBKey());
     }
-    static PassRefPtr<IDBKey> create(int32_t number)
+    static PassRefPtr<IDBKey> create(double number)
     {
         return adoptRef(new IDBKey(number));
     }
@@ -68,7 +68,7 @@ public:
         return m_string;
     }
 
-    int32_t number() const
+    double number() const
     {
         ASSERT(m_type == NumberType);
         return m_number;
@@ -88,12 +88,12 @@ public:
 
 private:
     IDBKey();
-    explicit IDBKey(int32_t);
+    explicit IDBKey(double);
     explicit IDBKey(const String&);
 
     Type m_type;
     String m_string;
-    int32_t m_number;
+    double m_number;
 };
 
 }
